backend/users/pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = ""
authors = ["Bogdan Loginov <bogdanloginov31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
django-admin-interface = "^0.28.6"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
